Remove ephemera old reward version
